Accessibility Change
I thiiiiink this should introduce a very rudimentary accessibility for blind players to use Monarch. Adds each card's Play button to the tab index, allowing the player to tab through their hand one at a time. Also adds an aria label that names the card and indicates that the button may be pressed to play it.